+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ## 0.2.0 (TBD)
 - Support variable watches
 - Support debug console command completion
+- Support disassembly view
+- Show registers in variables pane
 - Correctly escape debug console commands containing `"`
 - Upgrade dependencies
 - Other minor stability improvements